Refactor: move classes to persistence package
