+++ p2p/src/main/java/bisq/network/p2p/storage/persistence/RemovedPayloadsMap.java
@@ -15,8 +15,10 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.network.p2p.storage;
+package bisq.network.p2p.storage.persistence;
 
+
+import bisq.network.p2p.storage.P2PDataStorage;
 
 import bisq.common.proto.persistable.PersistableEnvelope;
 import bisq.common.util.Utilities;
